feat(construct.awscdk.github-pipeline): extract pipeline construct from `stacks.api`
