+++ packages/construct/awscdk/github-pipeline/.gitattributes
@@ -1,0 +1,21 @@
+# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+
+/.eslintrc.json linguist-generated
+/.gitattributes linguist-generated
+/.gitignore linguist-generated
+/.npmignore linguist-generated
+/.npmrc linguist-generated
+/.prettierignore linguist-generated
+/.prettierrc.json linguist-generated
+/.projen/** linguist-generated
+/.projen/deps.json linguist-generated
+/.projen/files.json linguist-generated
+/.projen/tasks.json linguist-generated
+/API.md linguist-generated
+/build.config.ts linguist-generated
+/LICENSE linguist-generated
+/package.json linguist-generated
+/pnpm-lock.yaml linguist-generated
+/project.json linguist-generated
+/tsconfig.dev.json linguist-generated
+/tsconfig.json linguist-generated
