+++ packages/construct/awscdk/github-pipeline/.gitignore
@@ -1,0 +1,47 @@
+# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+!/.gitattributes
+!/.projen/tasks.json
+!/.projen/deps.json
+!/.projen/files.json
+!/package.json
+!/LICENSE
+!/.npmignore
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+lerna-debug.log*
+report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
+pids
+*.pid
+*.seed
+*.pid.lock
+lib-cov
+coverage
+*.lcov
+.nyc_output
+build/Release
+node_modules/
+jspm_packages/
+*.tsbuildinfo
+.eslintcache
+*.tgz
+.yarn-integrity
+.cache
+!/.projenrc.js
+!/.prettierignore
+!/.prettierrc.json
+!/.npmrc
+!/test/
+!/src/
+/dist
+/dist/
+!/.eslintrc.json
+.jsii
+tsconfig.json
+!/API.md
+!/tsconfig.json
+!/tsconfig.dev.json
+!/build.config.ts
+!/project.json
