+++ packages/construct/awscdk/github-pipeline/.npmignore
@@ -1,0 +1,17 @@
+# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+/.projen/
+/test/
+/src/
+!/dist/
+!/dist/**/*.js
+!/dist/**/*.d.ts
+dist
+/.github/
+/.vscode/
+/.idea/
+/.projenrc.js
+tsconfig.tsbuildinfo
+/.eslintrc.json
+!.jsii
+/tsconfig.json
+/tsconfig.dev.json
