+++ packages/construct/awscdk/github-pipeline/.prettierignore
@@ -1,0 +1,13 @@
+# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+.eslintrc.json
+.gitattributes
+.gitignore
+.prettierignore
+.prettierrc.json
+.projen/deps.json
+.projen/files.json
+.projen/tasks.json
+LICENSE
+project.json
+tsconfig.dev.json
+tsconfig.json
