+++ packages/construct/awscdk/github-pipeline/.prettierrc.json
@@ -1,0 +1,7 @@
+{
+  "overrides": [],
+  "singleQuote": true,
+  "semi": false,
+  "useTabs": true,
+  "tabWidth": 2
+}
