+++ packages/construct/awscdk/github-pipeline/build.config.ts
@@ -1,0 +1,19 @@
+// ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+
+import { defineBuildConfig } from 'unbuild'
+
+export default defineBuildConfig({
+	name: 'construct.awscdk.github-pipeline',
+	declaration: true,
+	clean: true,
+	entries: ['./src/index'],
+	sourcemap: true,
+	rollup: {
+		emitCJS: true,
+		cjsBridge: true,
+		esbuild: {
+			treeShaking: true,
+			sourcemap: true,
+		},
+	},
+})
