+++ packages/construct/awscdk/github-pipeline/package.json
@@ -1,23 +1,26 @@
 {
-  "name": "@crisiscleanup/stacks.api",
+  "name": "@crisiscleanup/construct.awscdk.github-pipeline",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/CrisisCleanup/infrastructure"
+  },
   "scripts": {
     "build": "npx projen build",
-    "bundle": "npx projen bundle",
     "clean": "pnpm exec projen clean",
+    "compat": "npx projen compat",
     "compile": "npx projen compile",
     "default": "npx projen default",
-    "deploy": "npx projen deploy",
-    "destroy": "npx projen destroy",
-    "diff": "npx projen diff",
+    "docgen": "npx projen docgen",
     "eslint": "npx projen eslint",
     "package": "npx projen package",
+    "package-all": "npx projen package-all",
+    "package:js": "npx projen package:js",
     "post-compile": "npx projen post-compile",
+    "post-install": "npx projen post-install",
     "post-upgrade": "npx projen post-upgrade",
     "pre-compile": "npx projen pre-compile",
-    "synth": "npx projen synth",
-    "synth:silent": "npx projen synth:silent",
+    "stub": "npx projen stub",
     "test": "npx projen test",
-    "test:watch": "npx projen test:watch",
     "upgrade": "npx projen upgrade",
     "watch": "npx projen watch"
   },
@@ -31,50 +34,65 @@
     "@types/node": "^18",
     "@typescript-eslint/eslint-plugin": "^5",
     "@typescript-eslint/parser": "^5",
-    "aws-cdk": "^2.91.0",
-    "esbuild": "^0.19.1",
+    "aws-cdk-lib": "2.91.0",
+    "constructs": "10.2.69",
     "eslint": "^8",
     "eslint-config-prettier": "^8.10.0",
     "eslint-import-resolver-node": "^0.3.9",
     "eslint-import-resolver-typescript": "^3.6.0",
     "eslint-plugin-import": "^2.28.0",
     "eslint-plugin-prettier": "^5.0.0",
+    "jsii": "~5",
+    "jsii-diff": "^1.87.0",
+    "jsii-docgen": "^9.1.2",
+    "jsii-pacmak": "^1.87.0",
+    "jsii-rosetta": "~5",
     "lint-staged": "^13.2.3",
     "npm-check-updates": "^16",
     "prettier": "^3.0.1",
     "projen": "^0.72.11",
-    "ts-node": "^10.9.1",
     "typescript": "~5.1",
-    "vitest": "^0.34.1"
+    "unbuild": "^2.0.0-rc.0"
+  },
+  "peerDependencies": {
+    "aws-cdk-lib": "^2.91.0",
+    "constructs": "^10.2.69"
   },
   "dependencies": {
-    "@aws-cdk/lambda-layer-kubectl-v27": "^2.0.0",
-    "@aws-quickstart/eks-blueprints": "^1.10.1",
-    "@cdklabs/cdk-validator-cfnguard": "^0.0.54",
-    "@crisiscleanup/charts.crisiscleanup": "workspace:*",
     "@crisiscleanup/config": "workspace:*",
-    "@crisiscleanup/construct.awscdk.github-pipeline": "workspace:*",
-    "@crisiscleanup/k8s.construct.api": "workspace:*",
-    "@crisiscleanup/k8s.construct.component": "workspace:*",
-    "@kubecost/kubecost-eks-blueprints-addon": "^0.1.8",
-    "aws-cdk-lib": "^2.91.0",
-    "cdk-ec2-key-pair": "^3.3.1",
     "cdk-pipelines-github": "^0.4.103",
-    "cdk-sops-secrets": "^1.2.250",
-    "cdk8s": "2.42.0",
-    "cdk8s-plus-27": "2.6.0",
-    "constructs": "^10.2.69",
-    "defu": "^6.1.2",
-    "flat": "^5.0.2",
-    "zod": "^3.21.4"
+    "flat": "^5.0.2"
   },
+  "keywords": [
+    "cdk"
+  ],
   "engines": {
     "node": ">= 18.16.0"
   },
   "license": "Apache-2.0",
   "version": "0.0.0",
+  "types": "./dist/index.d.ts",
+  "stability": "stable",
+  "jsii": {
+    "outdir": "dist",
+    "targets": {},
+    "tsc": {
+      "outDir": "dist",
+      "rootDir": "src"
+    }
+  },
   "type": "module",
   "sideEffects": true,
+  "exports": {
+    ".": {
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.cjs"
+    },
+    "./package.json": "./package.json"
+  },
+  "files": [
+    "dist"
+  ],
   "packageManager": "pnpm@8.6.9",
   "lint-staged": {
     "*.{ts,tsx,mts,cts,js,jsx,jts,jts}": [
