+++ packages/construct/awscdk/github-pipeline/project.json
@@ -1,0 +1,148 @@
+{
+  "name": "construct.awscdk.github-pipeline",
+  "root": "packages/construct/awscdk/github-pipeline",
+  "targets": {
+    "build": {
+      "inputs": [
+        "default",
+        "^default",
+        "!{projectRoot}/.jsii",
+        "!{projectRoot}/dist/**/*",
+        "!{projectRoot}/dist/**/*"
+      ],
+      "outputs": [
+        "{projectRoot}/.jsii",
+        "{projectRoot}/dist",
+        "{projectRoot}/dist"
+      ],
+      "dependsOn": [
+        "^build"
+      ],
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen build",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "default": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen default",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "pre-compile": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen pre-compile",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "compile": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen compile",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "post-compile": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen post-compile",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "test": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen test",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "package": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen package",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "post-upgrade": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen post-upgrade",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "upgrade": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen upgrade",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "watch": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen watch",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "compat": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen compat",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "package-all": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen package-all",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "package:js": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen package:js",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "docgen": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen docgen",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "eslint": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen eslint",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "stub": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen stub",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "post-install": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen post-install",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    },
+    "clean": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "pnpm exec projen clean",
+        "cwd": "packages/construct/awscdk/github-pipeline"
+      }
+    }
+  },
+  "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
+}
