+++ packages/construct/awscdk/github-pipeline/tsconfig.dev.json
@@ -1,0 +1,16 @@
+// ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {},
+  "include": [
+    ".projenrc.js",
+    "src/**/*.ts",
+    "test/**/*.ts",
+    "src/*.ts",
+    "*.ts",
+    "**/*.ts"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
+}
