+++ packages/construct/awscdk/github-pipeline/tsconfig.json
@@ -1,0 +1,30 @@
+// ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+{
+  "extends": [
+    "../../../../tsconfig/tsconfig.base.json",
+    "../../../../tsconfig/tsconfig.esm.json",
+    "../../../../tsconfig/tsconfig.bundler.json",
+    "../../../../tsconfig/tsconfig.composite.json"
+  ],
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "dist",
+    "declarationDir": "src"
+  },
+  "include": [
+    "src/*.ts",
+    "src/**/*.ts"
+  ],
+  "exclude": [],
+  "typedocOptions": {
+    "entryPoints": [
+      "src/index.ts"
+    ],
+    "useTsLinkResolution": true
+  },
+  "references": [
+    {
+      "path": "../../../config/tsconfig.json"
+    }
+  ]
+}
